9th commit
js file were splitede exported and imported in javascript.js.
Now only the year appears on movie cards.
