+++ src/index.html
@@ -13,19 +13,24 @@
       rel="stylesheet"
     />
     <!-- Styles CSS -->
-    <link rel="stylesheet" href="../node_modules/notiflix/dist/notiflix-3.2.6.min.css">
-    <link rel="stylesheet" href="../node_modules/simplelightbox/dist/simple-lightbox.min.css">
+    <link
+      rel="stylesheet"
+      href="../node_modules/notiflix/dist/notiflix-3.2.6.min.css"
+    />
+    <link
+      rel="stylesheet"
+      href="../node_modules/simplelightbox/dist/simple-lightbox.min.css"
+    />
     <link rel="stylesheet" href="./css/modern-normalize.css" />
     <link rel="stylesheet" href="./css/header.css" />
   </head>
   <body>
+    <!-- HTML -->
     <include src="./partials/header.html"></include>
     <include src="./partials/spinner.html"></include>
-    <script type="module" src="./java-script/first-header.js"></script>
+    <!-- JavaScript -->
     <script src="../node_modules/notiflix/dist/notiflix-3.2.6.min.js"></script>
     <script src="../node_modules/axios/dist/axios.min.js"></script>
-    <script src="../node_modules/simplelightbox/dist/simple-lightbox.min.js"></script>
-    <script type="module" src="./java-script/second-header.js"></script>
-    <script type="module" src="./java-script/spinner.js"></script>
+    <script type="module" src="./index.js"></script>
   </body>
 </html>
