+++ src/index.js
@@ -1,0 +1,7 @@
+import './javasscript/api.js';
+import './javasscript/header-bkg.js'
+import './javasscript/letters-jump.js';
+import './javasscript/main.js';
+import './javasscript/notification.js';
+import './javasscript/spinner.js';
+import './javasscript/up-dwn-btn.js';
