+++ src/javasscript/notification.js
@@ -1,0 +1,5 @@
+import Notiflix from 'notiflix';
+
+export function showFailureNotification(message) {
+  Notiflix.Notify.failure(message);
+}
